Refactor: MemberVO class_no 추가, MemberDTO 변경 및 getlist 쿼리문 수정 Design:마이페이지, 관리자페이지 수정

+++ gooddev_web/src/main/java/io/good/gooddev_web/search/dto/PageResponseDTO.java
@@ -11,7 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 @Setter
 public class PageResponseDTO<T> {
     private int page;
-	private double size;
+	private int size;
 	private int total;
 	
 	private int start;
@@ -27,9 +27,9 @@ public class PageResponseDTO<T> {
 		this.page = pageRequest.getPage();
 		this.size = pageRequest.getSize();
 		this.total = total;
-		this.end = (int) (Math.ceil(page/size)*size);
+		this.end = (int) (Math.ceil(page/(double)size)*size);
 		this.start = end-((int)size-1);
-		int last = (int)Math.ceil(total/size);
+		int last = (int)Math.ceil(total/(double)size);
 		this.end = end > last ? last : end;
 		this.prev = this.start > 1;
 		this.next = total>this.end*this.size;
